\name{meshblocknz-package}
\alias{meshblocknz-package}
\docType{package}
\title{
MeshblocksNZ
}
\description{
Useful functions and data for dealing with spatial meshblocks in NZ
}
\details{
Contains data on meshblocks in NZ and (in future) functions for extracting features.
The key data is mb_2006_ur.
}
\author{
Jonathan Marshall
}
\references{
http://www.stats.govt.nz/Census/2006CensusHomePage/MeshblockDataset.aspx
}
\keyword{ package }
\seealso{
}
\examples{
head(mb_2006_ur)
}
