% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshblocknz.R
\docType{data}
\name{mb2006}
\alias{mb2006}
\title{Meshblock summary information for NZ from 2006}
\format{A data frame with 41376 rows and 11 variables:
\describe{
  \item{MB2006}{Meshblock identifier from census 2006.}
  \item{UR2006_num}{Numeric scale of urban/rural status from -3 (most rural) to 3 (most urban), based on 2006 meshblocks.}
  \item{UR2006}{Urban/Rural categorisation with 2 levels ("Rural" and "Urban") from 2006. Equivalent to UR2006_num <= 0}
  \item{DHB_code}{District Health Board code for this meshblock, from 2013 census areas.}
  \item{DHB_name}{District Health Board name for this meshblock, from 2013 census areas (Factor with 21 levels).}
  \item{AU2006}{Area unit code for this meshblock, from 2006 census areas.}
  \item{AU2006_name}{Area unit name for this meshblock, from 2006 census areas.}
  \item{TA2006}{Territorial Authority code for this meshblock, from 2006 census areas.}
  \item{TA2006_name}{Area unit name for this meshblock, from 2006 census areas.}
  \item{Ward2006}{Ward code for this meshblock, from 2006 census areas.}
  \item{Ward2006_name}{Ward name for this meshblock, from 2006 census areas.}
  \item{RC2006}{Regional Council code for this meshblock, from 2006 census areas.}
  \item{RC2006_name}{Regional Council name for this meshblock, from 2006 census areas.}
  \item{Pop2001}{Population of meshblock in 2001.}
  \item{Pop2006}{Population of meshblock in 2006.}
  \item{Pop2013}{Population of meshblock in 2013.}
}}
\source{
\url{http://www.stats.govt.nz/Census/2013-census.aspx}
}
\usage{
mb2006
}
\description{
A dataset containing the meshblock 2006 identifier and some summary information.
This is essentially a filtered version of `mb2013`, with population information
summed where meshblocks have been split from 2006 to 2013.
}
\keyword{datasets}

