% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshblocknz.R
\docType{data}
\name{mb2013}
\alias{mb2013}
\title{Meshblock summary information for NZ from 2013}
\format{A data frame with 46621 rows and 11 variables:
\describe{
  \item{MB2013}{Meshblock identifier from census 2013}
  \item{MB2006}{Meshblock identifier from census 2006. Contains duplicates.}
  \item{UR2013_num}{Numeric scale of urban/rural status from -1 (most rural) to 3 (most urban), based on 2013 meshblocks. Note, not the same scale as the 2006 classification, but <=0 is the same.}
  \item{UR2013_name}{Named urban/rural status, based on 2013 meshblocks.}
  \item{UR2006_num}{Numeric scale of urban/rural status from -3 (most rural) to 3 (most urban), based on 2006 meshblocks.}
  \item{UR2006_name}{Named urban/rural status, based on 2006 meshblocks.}
  \item{UR2013}{Urban/Rural categorisation with 2 levels ("Rural" and "Urban") from 2013. Equivalent to UR2013_num <= 0}
  \item{UR2006}{Urban/Rural categorisation with 2 levels ("Rural" and "Urban") from 2006. Equivalent to UR2006_num <= 0}
  \item{DHB_code}{District Health Board code for this meshblock, from 2013 census areas.}
  \item{DHB_name}{District Health Board name for this meshblock, from 2013 census areas (Factor with 21 levels).}
  \item{AU2013}{Area unit code for this meshblock, from 2013 census areas.}
  \item{AU2013_name}{Area unit name for this meshblock, from 2013 census areas.}
  \item{TA2013}{Territorial Authority code for this meshblock, from 2013 census areas.}
  \item{TA2013_name}{Area unit name for this meshblock, from 2013 census areas.}
  \item{Ward2013}{Ward code for this meshblock, from 2013 census areas.}
  \item{Ward2013_name}{Ward name for this meshblock, from 2013 census areas.}
  \item{CB2013}{Community Board code for this meshblock, from 2013 census areas.}
  \item{CB2013_name}{Community Board name for this meshblock, from 2013 census areas.}
  \item{RC2013}{Regional Council code for this meshblock, from 2013 census areas.}
  \item{RC2013_name}{Regional Council name for this meshblock, from 2013 census areas.}
  \item{GC2013}{General Constituency code for this meshblock, from 2013 census areas.}
  \item{GC2013_name}{General Constituency name for this meshblock, from 2013 census areas.}
  \item{MC2013}{M훮ori Constituency code for this meshblock, from 2013 census areas.}
  \item{MC2013_name}{M훮ori Constituency name for this meshblock, from 2013 census areas.}
  \item{GE2007}{General Electorate code for this meshblock, from 2007 election areas.}
  \item{GE2007_name}{General Electorate name for this meshblock, from 2007 election areas.}
  \item{ME2007}{M훮ori Electorate code for this meshblock, from 2007 election areas.}
  \item{ME2007_name}{M훮ori Electorate name for this meshblock, from 2007 election areas.}
  \item{LC2013}{Land Code for this meshblock, from 2013 census areas.}
  \item{LC2013_name}{Land Code name for this meshblock, from 2013 census areas.}
  \item{Area}{Area of this meshblock in square kilometers.}
  \item{Pop2001}{Population of meshblock in 2001.}
  \item{Pop2006}{Population of meshblock in 2006.}
  \item{Pop2013}{Population of meshblock in 2013.}
}}
\source{
\url{http://www.stats.govt.nz/Census/2013-census.aspx}
}
\usage{
mb2013
}
\description{
A dataset containing the meshblock 2013 identifier and some summary information
}
\keyword{datasets}
